requirements.txt
# Production Menu Extractor - Requirements
google-cloud-vision>=3.4.0
flask>=2.3.0
flask-limiter>=3.5.0
pillow>=10.0.0
flask==2.3.3
google-cloud-vision==3.4.4
Pillow==10.0.1
gunicorn==21.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
locust>=2.16.0

# Production Deployment
gunicorn>=21.2.0
redis>=4.6.0

# Optional: Enhanced OCR alternatives
# azure-cognitiveservices-vision-computervision>=0.9.0
# boto3>=1.28.0  # For AWS Textract